Move wrapper main class to proper package
Reason: Consistency with already implemented wrappers
