+++ minecraft/1.8/src/main/java/nova/worldgen/wrapper/mc/forge/v18/launch/NovaWorldgenWrapper.java
@@ -18,37 +18,36 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.worldgen.wrapper.mc.forge.v1_11;
+package nova.worldgen.wrapper.mc.forge.v18.launch;
 
 import net.minecraftforge.fml.common.registry.GameRegistry;
 import nova.core.loader.Loadable;
 import nova.core.loader.Mod;
 import nova.worldgen.WorldgenManager;
-import nova.worldgen.wrapper.mc.forge.v1_11.wrapper.world.forward.FWWorldGenerator;
+import nova.worldgen.wrapper.mc.forge.v18.wrapper.world.forward.FWWorldGenerator;
 
 /**
  *
  * @author ExE Boss
  */
-@Mod(id = NovaMinecraftWorldgen.id, name = NovaMinecraftWorldgen.name, version = NovaMinecraftWorldgen.version, novaVersion = "0.0.1")
-public class NovaMinecraftWorldgen implements Loadable {
+@Mod(id = NovaWorldgenWrapper.id, name = NovaWorldgenWrapper.name, version = NovaWorldgenWrapper.version, novaVersion = "0.0.1")
+public class NovaWorldgenWrapper implements Loadable {
 
 	public static final String version = "0.0.1";
-	public static final String id = "novaworldgen";
+	public static final String id = "nova-worldgen-wrapper";
 	public static final String name = "NOVA Worldgen";
 
 	public final WorldgenManager worldgenManager;
 
 	public FWWorldGenerator worldGenerator;
 
-	public NovaMinecraftWorldgen(WorldgenManager worldgenManager) {
+	public NovaWorldgenWrapper(WorldgenManager worldgenManager) {
 		this.worldgenManager = worldgenManager;
-
-		worldGenerator = new FWWorldGenerator(worldgenManager);
 	}
 
 	@Override
-	public void preInit() {
+	public void init() {
+		worldGenerator = new FWWorldGenerator(worldgenManager);
 		GameRegistry.registerWorldGenerator(worldGenerator, 0);
 	}
 }
